Use gc methods to collect functions to update (#4)
This implements the method discussed in #3, following suggestions by @alexmojaki. This allows jurigged to do more thorough updates, chiefly when closures are involved, but also when functions are stashed by decorators into data structures.

Essentially, `gc.get_objects()` is used to get an initial inventory of code objects, `sys.audithook` is used to complete the inventory as `exec` is called, and `gc.get_referrers` is used to map code objects to functions. Since the latter method gets expensive as more objects are created on the heap, the results are cached and the cache is used once the method gets slower than 0.1s.